Update BNCStrongMatchHelper to handle UISplitViewController (#625)
Similar to how UINavigationController and UITabBarController is
currently being handled when adding BNCMatchView. If we don't handle
UISplitViewController this way, BNCMatchViewController will end up in
the split view's viewControllers property which is unexpected and may
lead to crashes later. Split view's viewControllers should only have 1
or 2 view controllers depending on its collapsed / expanded state.

The special handling of UINavigationController and UITabBarController
was added in
https://github.com/BranchMetrics/ios-branch-deep-linking/pull/539
Former-commit-id: 6fc35fcdcabe1747e090281acc2deefb90f58888 [formerly dd5ccb68f5f92351170487e2d66c2c2ec2965504]
Former-commit-id: cc192dd4c5b2800cff4f9bd733bf58bc58d47c22